--- Angular/fronend-analysis/src/styles.scss
@@ -1,1 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
