+++ backend/src/criteria/base_criterion.py
