+++ backend/src/criteria/function_Extractor.py
