+++ backend/tests/conftest.py
@@ -1,0 +1,14 @@
+import pytest
+from src.backend import app
+
+@pytest.fixture(scope="module")
+def client():
+    app.config.update({
+        "TESTING": True,
+    })
+
+    yield app.test_client()
+
+    app.config.update({
+        "TESTING": False,
+    })
