+++ backend/tests/test_analyzer.py
@@ -1,0 +1,4 @@
+class TestAnalyzer():
+    def test_insert_empty_directory_returns_error(self):
+        assert True
+
