+++ frontend/fronend-analysis/public/favicon.ico
