+++ frontend/fronend-analysis/src/app/app.component.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <h1>{{ title }}</h1>
+  <h1>Code Analysis Tool</h1>
   <label for="repoUrl">Git Repository URL: </label>
   <input
     id="repoUrl"
