+++ frontend/fronend-analysis/src/app/app.component.scss
