+++ frontend/fronend-analysis/src/app/app.component.spec.ts
