+++ frontend/fronend-analysis/tsconfig.spec.json
